Refactor: move header to layout of (top) - header displays different titles according to the pathname
part of SWM-102
related to #27
