+++ src/app/(top)/workbookDetail/[workbookId]/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useParams } from 'next/navigation';
-import Header from '@/app/(top)/workbookDetail/[workbookId]/_components/header';
+import Header from '@/app/(top)/_components/header';
 import WorkbookInformation from '@/app/(top)/workbookDetail/[workbookId]/_components/workbook-information';
 import WorkbookEditButtons from '@/app/(top)/workbookDetail/[workbookId]/_components/workbook-edit-buttons';
 import WorkbookQuestionArea from '@/app/(top)/workbookDetail/[workbookId]/_components/workbook-question-area';
@@ -11,7 +11,6 @@ const Page = () => {
 
   return (
     <div>
-      <Header />
       <WorkbookInformation />
       <WorkbookEditButtons />
       <WorkbookQuestionArea />
